% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFPois.R
\name{gkl.dev}
\alias{gkl.dev}
\title{Function for calculating the generalised Kullback Leibler divergence}
\usage{
gkl.dev(y, mu)
}
\arguments{
\item{y}{Observation}

\item{mu}{Estimate}
}
\value{
Generalized Kullback-Leibler
}
\description{
Internal function used in NMFPois
}
