% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFPois.R
\name{NMFPois}
\alias{NMFPois}
\title{Non-negative matrix factorization algorithm for Poisson data

Factorizing M into two matrices P and E of
dimension ncol(M) x N and N x nrow(M) with the acceleration of SQUAREM.
The objective function is the generalized Kullback-Leibler divergence(GKLD).}
\usage{
NMFPois(M, N, seed = sample(1:100, 3), arrange = TRUE, tol = 1e-05)
}
\arguments{
\item{M}{Non-negative data matrix of size}

\item{N}{Small dimension of the two new matrices}

\item{seed}{Vector of random seeds to initialize the matrices}

\item{arrange}{Arranging columns in P and rows in E after largest row sums of E}

\item{tol}{Maximum change of P and E when stopping}
}
\value{
A list of the matrices derived by the factorization and the corresponding generalized Kullback-Leibler
 \itemize{
 \item P   - Non-negative matrix of dimension ncol(V) x K, with columns summing to one
 \item E   - Non-negative matrix of dimension K x nrow(V), where rows sum to one 
 \item gkl - Smallest Value of the Generalized Kullback-Leibler
 }
}
\description{
Non-negative matrix factorization algorithm for Poisson data

Factorizing M into two matrices P and E of
dimension ncol(M) x N and N x nrow(M) with the acceleration of SQUAREM.
The objective function is the generalized Kullback-Leibler divergence(GKLD).
}
