% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSFS.R
\name{plotSFS}
\alias{plotSFS}
\title{Plotting the set of feasible solutions}
\usage{
plotSFS(sampleSFSOutput)
}
\arguments{
\item{sampleSFSOutput}{The output from the sampleSFS function}
}
\value{
Plot of the SFS for P and E
}
\description{
Plotting the set of feasible solutions
}
