% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleSFS}
\alias{sampleSFS}
\title{Finds the SFS from a given solution of P and E}
\usage{
sampleSFS(P, E, maxIter = 10^5L, check = 1000L, beta = 0.5, eps = 1e-05)
}
\arguments{
\item{P}{One of the factorized two matrices, which will control the stopping criteria}

\item{E}{The other of the facotrized matrizes}

\item{maxIter}{The maximum number of iterations}

\item{check}{Number of iterations between checking to stop}

\item{beta}{The two shape parameters in the be3ta distribution to sample lambda}

\item{eps}{Epsilon for the stopping criteria}
}
\value{
A list of outputs from the sampleSFS function
\itemize{
        \item avgChangeFinal     - final average change for each entrance in P
        \item avgChangeEachCheck - average change for each entrance in P at each check
        \item totalIter          - total number of iterations needed for convergence of the average change
        \item Pminimum           - minimum for each entry in P
        \item Pmaximum           - maximum for each entry in P
        \item Eminimum           - minimum for each entry in E
        \item Emaximum           - maximum for each entry in E
        \item P_lastCheckResults - results of P for the last 'check' iterations 
        \item E_lastCheckResults - results of E for the last 'check' iterations 
        }
}
\description{
Finds the SFS from a given solution of P and E
}
