\name{SFS-package}
\alias{SFS-package}
\alias{SFS}
\docType{package}
\title{
\packageTitle{SFS}
}
\description{
\packageDescription{SFS}
}
\details{

\packageDESCRIPTION{SFS}
\packageIndices{SFS}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{SFS}

Maintainer: \packageMaintainer{SFS}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\keyword{ cluster }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
