\name{SFS-package}
\alias{SFS-package}
\alias{SFS}
\docType{package}
\title{
\packageTitle{SFS}
}
\description{
\packageDescription{SFS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SFS}
\packageIndices{SFS}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{SFS}

Maintainer: \packageMaintainer{SFS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
