\name{SFS-package}
\alias{SFS-package}
\alias{SFS}
\docType{package}
\title{
\packageTitle{SFS}
}
\description{
\packageDescription{SFS}
}
\details{

}
\author{
\packageAuthor{SFS}
Maintainer: \packageMaintainer{SFS}
}

\references{ 
M. Laurent and M. Seminaroti. Similarity-First Search: a new algorithm with application to Robinsonian matrix recognition. \emph{arXiv:1601.03521}. 2016.
    
M. Seminaroti. Combinatorial algorithms for the seriation problem. \emph{PhD thesis}. Center for Economic Research (CentER), Tilburg University, pages 1--209. 2016.
}

%\keyword{ package }
%\keyword{ cluster }
%\seealso{
%  \code{\link[seriation]{seriate}}
%  \code{\link{seriate.matrix}}
%  \code{\link{seriate.dist}}
%  \code{\link{dist}}
%}